concatenate string with n characters from src

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -9,13 +9,13 @@
  */
 char *_strncat(char *dest, char *src, int n)
 {
-	int i = 0, j = 0;
+	unsigned int i = 0, j = 0;
 
 	while (*(dest + i) != '\0')
 		i++;
 	while (*(src + j) != '\0')
 	{
-		if (j < n)
+		if (j < (unsigned int) n)
 		{
 			*(dest + i) = *(src + j);
 			j++;
@@ -26,7 +26,7 @@ char *_strncat(char *dest, char *src, int n)
 			break;
 		}
 	}
-	if (j < n)
+	if (j < (unsigned int) n)
 		*(dest + i) = '\0';
 	return (dest);
 }
